BZ#1381425: Added info on Tenant Mapping
This commit adds an explanation of how Tenant Mapping works
by mapping OpenStack tenant lists to cloud tenants in Cloudforms/
ManageIQ.

New content added:

- Short explanation of how tenant mapping works
- Extra step for enabling tenant mapping when adding an OpenStack
  cloud provider

The content for that short explanation is single-sourced between a
subsection on the feature in Tenants (in the
doc-Managing_Infrastructure_and_Inventory doc) and a note in the
instructions for adding an OpenStack cloud provider (in the
doc-Managing_Providers doc).

I also took this opportunity to single-source an explanation of
Keystone V3 in the instructions for adding an OpenStack cloud
and infrastructure provider.