Update Policy creation 'towhat' field documentation
Original bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1659899

Messaging bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1663562

With the above bug, it was determined that the documentation - which
was used to test the Policy create action in the API - for the `towhat`
value was not reflective of a valid value for that field.

This change updates the documentation to reflect recent changes which
added an inclusion whitelist for `towhat` values on creation of Policies.

Relevant PRs:
Inclusion list: https://github.com/ManageIQ/manageiq/pull/18032/files
Error messaging: https://github.com/ManageIQ/manageiq/pull/18338