Update the OCP Add Container Docs for Alerts/Metrics
The documentation needs to be updated after migrating the new container
provider screen to DDF, specifically the "Detect" route button is no
longer available so we should document what `oc` command to run in order
to get the hostname.

Also we should indicate that Prometheus Alerts are not supported on
OCPv4.