Add EC2 Instance State-change CloudWatch Rule
The existing cloudwatch rules will raise an event when the API call to
stop an instance is issued but no event is raised when an instance
finishes shutting down.

The end result is the targeted refresh runs "too fast" and the instance
is "stuck" in powering_down.

We expect to have events raised for the "EC2 Instance State-change
Notification" group of events which raise when the instance state
changes to "stopped".

This CloudWatch rule appears to be missing from the docs for configuring
the amazon event catcher.