Drop usage of site.data.product.title_abbr
The only usage of this prodcutization string is the the azure install,
and it's only used in command line examples.  The actual values don't
require it, and generic strings can be used instead.  Additionally, it's
possible to productize with values that can create invalid command line
strings.  As such, this commit removes it entirely in favor of generic
strings.